6_sensorDataModel/requirements.txt
matplotlib==3.7.2
numpy==1.25.1
paho-mqtt==1.6.1
scikit-learn==1.3.0
scipy==1.11.1
